src/requirements.txt
streamlit
langchain
langgraph
openai
python-docx
pdfplumber

